Rearraned examples - removed Maven setup from examples. Need to add specific Maven setup example
