+++ tightdb-example/.settings/org.eclipse.jdt.apt.core.prefs
@@ -1,4 +1,3 @@
-#Wed Sep 12 00:03:18 CEST 2012
 eclipse.preferences.version=1
 org.eclipse.jdt.apt.aptEnabled=true
 org.eclipse.jdt.apt.genSrcDir=target/generated
