+++ tightdb-example/.settings/org.eclipse.jdt.core.prefs
@@ -1,4 +1,3 @@
-#Wed Sep 12 00:00:30 CEST 2012
 eclipse.preferences.version=1
 org.eclipse.jdt.core.compiler.codegen.inlineJsrBytecode=enabled
 org.eclipse.jdt.core.compiler.codegen.targetPlatform=1.6
