--- tightdb-example/.settings/org.maven.ide.eclipse.prefs
@@ -1,8 +1,0 @@
-#Sun May 13 04:42:10 CEST 2012
-activeProfiles=
-eclipse.preferences.version=1
-fullBuildGoals=process-test-resources
-resolveWorkspaceProjects=true
-resourceFilterGoals=process-resources resources\:testResources
-skipCompilerPlugin=true
-version=1
