--- tightdb-example/pom.xml
@@ -1,67 +1,0 @@
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-	xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-	<modelVersion>4.0.0</modelVersion>
-
-	<groupId>com.tightdb</groupId>
-	<artifactId>tightdb-example</artifactId>
-	<version>0.0.1-SNAPSHOT</version>
-
-	<properties>
-		<project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-		<project.reporting.outputEncoding>UTF-8</project.reporting.outputEncoding>
-	</properties>
-
-    <build>
-        <plugins>
-        	<plugin>
-            	<groupId>org.apache.maven.plugins</groupId>
-                <artifactId>maven-compiler-plugin</artifactId>
-				<version>2.3.2</version>
-                <configuration>
-                	<source>1.6</source>
-                    <target>1.6</target>
-                </configuration>
-            </plugin>
-            
-			<plugin>
-				<groupId>org.bsc.maven</groupId>
-				<artifactId>maven-processor-plugin</artifactId>
-				<version>2.0.5</version>
-				<executions>
- 					<execution>
-   					<goals>
-     					<goal>process</goal>
-   					</goals>
-   					<configuration>
-   					    <processors>
-    					    <processor>com.tightdb.generator.CodeGenProcessor</processor>
-   					    </processors>
-   					    <outputDirectory>${project.build.directory}/generated</outputDirectory>
-   					</configuration>
- 					</execution>
-				</executions>
-				<dependencies>
-					<dependency>
-						<groupId>com.tightdb</groupId>
-						<artifactId>tightdb-devkit</artifactId>
-						<version>1.1.0</version>
-					</dependency>
-   				</dependencies>
-         	</plugin>
-		</plugins>
-	</build>
-
-	<dependencies>
-		<dependency>
-			<groupId>com.tightdb</groupId>
-			<artifactId>tightdb</artifactId>
-			<version>1.1.0</version>
-		</dependency>
-		<dependency>
-			<groupId>com.almworks.sqlite4java</groupId>
-			<artifactId>sqlite4java</artifactId>
-			<version>0.282</version>
-		</dependency>
-	</dependencies>
-
-</project>
