+++ tightdb-example/src/main/java/com/tightdb/examples/performance/PerformanceBase.java
@@ -1,4 +1,4 @@
-package com.tightdb.performance;
+package com.tightdb.examples.performance;
 
 public abstract class PerformanceBase implements IPerformance {
