+++ tightdb-example/src/main/java/com/tightdb/examples/performance/Util.java
@@ -1,4 +1,4 @@
-package com.tightdb.performance;
+package com.tightdb.examples.performance;
 
 import java.util.Scanner;
